Allow Expression as Value.
